Separate files to have each components on its own view.
